+++ internal/assayparser/io.go
@@ -1,16 +1,14 @@
 package assayparser
 
 import "os"
+
 // import "strings"
 import "encoding/json"
 import "gopkg.in/yaml.v3"
 import "errors"
 
-
 var ioError error = errors.New("Error in the IO process. Check inputs!")
 var DefaultPathJSON string = "./out/outputJSON.json"
-
-
 
 func ConvertJson(a ValidAssay) ([]byte, error) {
 	result, err := json.Marshal(a)
@@ -19,8 +17,6 @@ func ConvertJson(a ValidAssay) ([]byte, error) {
 	}
 	return result, nil
 }
-
-
 
 func ExportJson(path string, js []byte) error {
 	if path == "" || js == nil {
@@ -33,8 +29,6 @@ func ExportJson(path string, js []byte) error {
 	return nil
 }
 
-
-
 func ConvertYaml(a ValidAssay) ([]byte, error) {
 	result, err := yaml.Marshal(a)
 	if err != nil {
@@ -42,8 +36,6 @@ func ConvertYaml(a ValidAssay) ([]byte, error) {
 	}
 	return result, nil
 }
-
-
 
 func ExportYaml(path string, ya []byte) error {
 	if path == "" || ya == nil {
@@ -56,8 +48,6 @@ func ExportYaml(path string, ya []byte) error {
 	return nil
 }
 
-
-
 func ImportFile(path string) ([]byte, error) {
 	data, err := os.ReadFile(path)
 	if err != nil {
@@ -66,20 +56,20 @@ func ImportFile(path string) ([]byte, error) {
 	return data, nil
 }
 
-
-
 func UnwindJson(in []byte) (ValidAssay, error) {
 	var a ValidAssay
 	err := json.Unmarshal(in, &a)
-	if err != nil {return ValidAssay{}, ioError;}
+	if err != nil {
+		return ValidAssay{}, ioError
+	}
 	return a, nil
 }
-
-
 
 func UnwindYaml(in []byte) (ValidAssay, error) {
 	var a ValidAssay
 	err := yaml.Unmarshal(in, &a)
-	if err != nil {return ValidAssay{}, ioError;}
+	if err != nil {
+		return ValidAssay{}, ioError
+	}
 	return a, nil
 }